adding the trigger for AvertageRating

+++ WebApplication2/ViewModels/AddCourseViewModel.cs
@@ -27,7 +27,7 @@ namespace WebApplication2.ViewModels
         public DateTime? AddingDate { get; set; } = DateTime.Now;
         public IFormFile? Picture { get; set; }
 
-        public int? AverageRating { get; set; } = 0;
+        public double? AverageRating { get; set; } = 0.0;
         public bool? Claimed { get; set; }
         public bool? Status { get; set; }
         //public string InstructorId { get; set; }
